move component directory
