+++ src/components/wallet-manager/index.test.tsx
