+++ src/components/wallet-manager/styles.scss
